Refactor HMAC implementation
