+++ src/lib.rs
@@ -7,13 +7,13 @@ mod curve25519;
 mod ed25519;
 mod hkdf_sha256;
 mod hkdf_sha512;
-mod hmac_sha256;
-mod hmac_sha512;
+mod hmac;
 mod poly1305;
 mod rng;
 mod sha256;
 mod sha512;
 mod sss;
+mod traits;
 mod verify;
 mod x25519;
 mod xor;
@@ -23,8 +23,9 @@ pub use crate::{
     ed25519::{ed25519_generate_key_pair, ed25519_public_key, ed25519_sign, ed25519_verify},
     hkdf_sha256::hkdf_sha256,
     hkdf_sha512::hkdf_sha512,
-    hmac_sha256::{hmac_sha256, hmac_sha256_verify, HmacSha256},
-    hmac_sha512::{hmac_sha512, hmac_sha512_verify, HmacSha512},
+    hmac::{
+        hmac_sha256, hmac_sha256_verify, hmac_sha512, hmac_sha512_verify, HmacSha256, HmacSha512,
+    },
     rng::Rng,
     sha256::{sha256, Sha256},
     sha512::{sha512, Sha512},
